Separate Feedback smart/dumb component
