+++ package.json
@@ -5,7 +5,7 @@
   "devDependencies": {
     "enzyme": "^2.4.1",
     "react-redux": "^4.4.5",
-    "react-scripts": "0.2.1",
+    "react-scripts": "0.4.3",
     "redux": "^3.6.0"
   },
   "dependencies": {
@@ -15,7 +15,8 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "test": "react-scripts test --env=jsdom"
   },
   "eslintConfig": {
     "extends": "./node_modules/react-scripts/config/eslint.js"
