+++ src/components/App.js
@@ -3,7 +3,7 @@ import Scoreboard from '../containers/scoreboard'
 import TimeTracker from '../containers/timetracker'
 import MainMenu from '../containers/mainmenu'
 import Calculator from '../containers/calculator'
-import Feedback from '../containers/feedback'
+import Feedback from './Feedback'
 
 const App = () => (
   <div className="App math-boggle">
