+++ src/containers/feedback.js
@@ -2,46 +2,7 @@ import React, {Component} from 'react'
 import {bindActionCreators} from 'redux'
 import {connect} from 'react-redux'
 import {updateFeedback} from '../actions/index'
-
-class Feedback extends Component {
-
-  constructor(){
-    super()
-    this.state = {
-      timeouts: []
-    }
-  }
-
-  componentWillReceiveProps(nextProps){
-    this.clearTimeouts()
-    let newTimeout = setTimeout(this.setVisibility, 1000)
-    this.setState({
-      timeouts: [newTimeout]
-    })
-  }
-
-  setVisibility = (visibility) => {
-    this.props.updateFeedback("")
-    this.clearTimeouts()
-  }
-
-  clearTimeouts = () => {
-    for(let timeout of this.state.timeouts) {
-      clearInterval(timeout)
-    }
-    this.setState({
-      timeouts: []
-    })
-  }
-
-  render(){
-    return (
-      <div className="feedback">
-        <div>{this.props.feedback}</div>
-      </div>
-    )
-  }
-}
+import Feedback from '../components/Feedback'
 
 function mapStateToProps(state){
   return {
